moved cli out, changed helpers
